==> 00_DotaApi.ipynb <==
# Dota API

The first step is to figure out how we can use the Dota API to get a bunch of historical games.



## Basic API Usage

We'll use this repository as a starting point: https://github.com/bryanveloso/dota2.py

It's a little old and unmaintained but so is the SteamAPI!

import os
import json
import time
import datetime
import requests

We need a Steam API Key. Go get it from: https://steamcommunity.com/dev/apikey

If you like living on the edge you can embed it directly in this notebook and pray that you never commit it to GitHub. Otherwise, export it from your `.bashrc` as `STEAM_API_KEY`.

if 'STEAM_API_KEY' not in os.environ:
  print("No API Key :(")
else:
  print("Found API Key.")
  STEAM_API_KEY = os.environ['STEAM_API_KEY']

base_url = 'https://api.steampowered.com'

def __request(method, path, **kwargs):
  url = base_url + path
  kwargs.setdefault('params', dict()).update(key=STEAM_API_KEY)
  response = requests.request(method, url, **kwargs)
  return response.json()

def get_steam_id(vanity_name, **params):
    path = '/ISteamUser/ResolveVanityURL/v0001'
    params.update(vanityurl=vanity_name)
    response = __request('get', path, params=params).get('response')
    if response and "success" in response and response['success'] == 1:
        return response
    else:
        print("Error:", response)

response = get_steam_id("Permit")
print("Response:", response)
steam_id = response['steamid']

Nice. So we can successfully hit the API and get my ID back. 

## Get Information About a Player

def get_player_summaries(steam_ids, **params):
  path = '/ISteamUser/GetPlayerSummaries/v0002'
  params.update(steamids=steam_ids)
  return __request('get', path, params=params)

response = get_player_summaries(steam_id)
print(response)

## Get Information About Games

Let's start by getting the most recent games available.

def get_most_recent_matches():
  path = '/IDOTA2Match_570/GetMatchHistory/V001'
  return __request('get', path)

response = get_most_recent_matches()
response

Great! It looks like the games are still in progress though. Some of them don't appear to have finished hero selection so they won't be very useful for us. 

Let's try getting older games.

def get_matches_before(match_id, **params):
  path = '/IDOTA2Match_570/GetMatchHistory/V001'
  params.update(start_at_match_id=match_id)
  return __request('get', path, params=params)

response = get_matches_before(5126114401)
response

Hmm this isn't working. Maybe it limits it to the most recent 500 games? That would suck. I will revisit this later.

Later: Yeah apparently this is essentially useless.

## Get Match History By Sequence Number

Here's another API endpoint that looks promising. Let's take a look.

def get_match_history_by_seq_num(seq_num, num_matches, **params):
  path = '/IDOTA2Match_570/GetMatchHistoryBySequenceNum/V001'
  params.update(start_at_match_seq_num=seq_num)
  params.update(matches_requested=num_matches)  
  return __request('get', path, params=params)

response = get_match_history_by_seq_num(5126114401, 1)
response

Great! This looks like what we'll need!

One last question: What is the maximum number of games we can retrieve at a time?

response = get_match_history_by_seq_num(5126114401, 1000)
len(response['result']['matches'])

Guess it's 100 at a time.
