==> 03_Baseline.ipynb <==
# Baseline Model

Before we make our real model we'd first like to create a baseline model. This model should be very simple and provide us with a baseline. If our real model does worth than this we'll know something is very wrong.

import os
import json
import time
import datetime
import requests
import numpy as np
import pandas as pd

df = pd.read_csv("training_set_small.csv")
df

# Create train/valid set
train = df.iloc[:9000]
valid = df.iloc[9000:]

len(train), len(valid)

min_hero = 1
max_hero = 135

num_games = [0 for _ in range(max_hero + 1)]
num_wins = [0 for _ in range(max_hero + 1)]

# Keep track of the numbers of games played for each hero and the number of times that hero won.
for i, row in train.iterrows():
    num_games[row['hero0']] += 1
    num_games[row['hero1']] += 1
    num_games[row['hero2']] += 1
    num_games[row['hero3']] += 1
    num_games[row['hero4']] += 1
    num_games[row['hero5']] += 1
    num_games[row['hero6']] += 1
    num_games[row['hero7']] += 1
    num_games[row['hero8']] += 1
    num_games[row['hero9']] += 1
    
    if row['radiant_win'] == True:
        num_wins[row['hero0']] += 1
        num_wins[row['hero1']] += 1
        num_wins[row['hero2']] += 1
        num_wins[row['hero3']] += 1
        num_wins[row['hero4']] += 1
    else:
        num_wins[row['hero5']] += 1
        num_wins[row['hero6']] += 1
        num_wins[row['hero7']] += 1
        num_wins[row['hero8']] += 1
        num_wins[row['hero9']] += 1

# Now create a lookup for each hero id that contains:
# - number of wins
# - number of games
# - win rate
lookup = pd.DataFrame(list(zip(num_wins, num_games)), 
                      columns=["num_wins", "num_games"], 
                      index=np.arange(max_hero + 1))
eps = 1e-6 # Avoid divide by zero.
lookup['win_rate'] = lookup['num_wins'] / (lookup['num_games'] + eps)

# For any heros we've never seen before, we'll assume a 50% win rate.
lookup[lookup['win_rate'] == 0.] = 0.5
lookup

Now that we have this lookup for each hero, we can generate predictions for new games (`valid`) based on which team has the highest aggregate win-rate.

valid

preds = []

for i, row in valid.iterrows():
    
    p_rad = ((
    lookup.iloc[row['hero0']]['win_rate'] + 
    lookup.iloc[row['hero1']]['win_rate'] + 
    lookup.iloc[row['hero2']]['win_rate'] + 
    lookup.iloc[row['hero3']]['win_rate'] + 
    lookup.iloc[row['hero4']]['win_rate']) / 5)
    
    p_dire = ((
    lookup.iloc[row['hero5']]['win_rate'] + 
    lookup.iloc[row['hero6']]['win_rate'] + 
    lookup.iloc[row['hero7']]['win_rate'] + 
    lookup.iloc[row['hero8']]['win_rate'] + 
    lookup.iloc[row['hero9']]['win_rate']) / 5)
    
    preds.append(p_rad >= p_dire)

valid['pred'] = preds


valid['correct'] = valid['pred'] == valid['radiant_win']

valid['correct'].value_counts()

counts = valid['correct'].value_counts()
counts[True] / len(valid)

We do slightly better than random!
