==> 06_CleanTrainingSet.ipynb <==
# Clean Training Set

Now that we've downloaded a bunch of Dota 2 games, let's clean up the dataset and format it so it's easier for us to train our model on it.

import os
import re
import json
import time
import tqdm
import datetime
import requests
import pandas as pd
from os.path import exists

# Stored outside of repo because it's 140GB
path = 'dota_2_bulk.tsv'

# How many games (lines) do we have?
num_games = sum(1 for line in open(path))
num_games

# Let's see what a single raw example looks like.
with open(path) as file:
  for line in tqdm.tqdm_notebook(file):
    seq_num, raw_json = line.split('\t')
    print(seq_num)
    print(raw_json)
    break

# Steam represents:
#  Players 0 - 4 as player slots 0 - 4
#  Players 5 - 9 as player slots 128-132
def mapPlayerSlotToPosition(slot):
  if slot <= 4:
    return slot
  return slot - 123

# Now read each line, parse the JSON and rip out the information we're interested in.
num_failures = 0
training_set = []

with open(path) as file:
  for line in tqdm.tqdm_notebook(file, total=num_games):
    seq_num, raw_json = line.split('\t')
    try:
      match = json.loads(raw_json)
      players = match['players']
      
      hero_ids = [None for _ in range(10)]
      account_ids = [None for _ in range(10)]
      
      if len(players) != 10:
        print("Weird. There weren't 10 players!")
        print(raw_json)
        num_failures += 1
        continue
      
      for player in players:
        player_slot = player['player_slot']
        hero_id = player['hero_id']
        # The default account number for anonymous accounts is 4294967295
        # Some games are missing the account number. I don't know why. 
        # We will default to the anonymous account number.
        account_id = 4294967295
        if 'account_id' in player:
          account_id = player['account_id']
          
        position = mapPlayerSlotToPosition(player_slot)
        hero_ids[position] = hero_id
        account_ids[position] = account_id
    
      start_time = match['start_time']
      # Convert to float target.
      radiant_win = float(match['radiant_win'])
      entry = hero_ids + account_ids + [start_time] + [radiant_win]
      training_set.append(entry)
      
    except Exception as e:
      print("Error:", e)
      print(raw_json)
      print(e)
      
      _, original = line.split('\t')
      print(original)
      num_failures +=1
      
print("Num failures:" , num_failures)

# Not sure why we're missing like 1 million rows...
len(training_set)

df = pd.DataFrame(training_set, 
             columns=['hero0', 'hero1', 'hero2', 'hero3', 'hero4',
                      'hero5', 'hero6', 'hero7', 'hero8', 'hero9',
                      'account0', 'account1', 'account2', 'account3', 'account4',
                      'account5', 'account6', 'account7', 'account8', 'account9',
                      'start_time', 'radiant_win'])
df.head()

df

df.to_csv("training_set_large.csv", index=False)
